Ajuste Formulário, quando for saque so podera escolher um type, somente Pix ou somente Wallet. E só exibe os campos de pix_key_type e pix_key, quando for saque do tipo Pix.
