+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "merchant-example-react-js",
-  "version": "2.0.3",
+  "version": "2.0.4",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.5.0",
