+++ src/components/Form/components/TypesTransaction/styles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const ContainerCheckTypes = styled.div`
+  width: 100%;
+  flex-direction: row;
+  height: auto;
+`;
